tested the cards
First, I created a separate package for the card classes and interfaces. Secondly, I needed a way of adding a card to a deck using the addCard() method, so I defined this method to take two arguments: the Suit and the Rank for a single card. This taught me how to use the Rank and Suit enum types which were defined inside the Card interface: I needed to import them in the class or interface where I was to use them. However, the challenge now is, the enitreDeck attribute is not instantiated with any values, it is null; and thus, a NullPointerException is returned. I need to solve this to proceed printing out the cards I have added to the deck.