printed cards in a deck
I solved the NullPointerException by instantiating the entireDeck attribute in StandardDeck using an ArrayList<Card> object since I could not instantiate it with an interface List. The first trial printed references to each card, but then modified the second attempt to get the suit and the rank using the defined methods.