multiple inheritance
Three classes Cuboid, Rectangle and Triangle implement the Irelatable interface with an abstract method which they all implement. Furthermore, they extend the class FindBySize, which has three methods these classes need. They all implement the abstract method according to the parameters which can be used to determine their sizes such as volume for the cuboids and area for the rectangles and triangles. These implementations of the abstract method is unique to each of them. In this way, these classes get functionality from a class and from an interface which is an example of multiple inheritance. Furthermore, multiple inheritance can be achieved when a class extends another and implements any number of interfaces.